feat: Implement RideMap component with optimized map rendering and location handling
- Added RideMap component to display routes and markers for rides.
- Introduced useDriverLocation hook to manage driver location updates.
- Refactored useGetCurrentLocation to utilize useLocationService for better location handling.
- Created useUserLocation hook for managing user location state.
- Implemented useLocationService for centralized location fetching logic.
- Updated driver and user navigation to utilize new hooks and components.
- Removed deprecated ActiveMap and RideMap pages, consolidating functionality into RideMap component.
- Enhanced location watcher with improved performance and logging.
- Updated styles for RideMap and added custom Google Map styles.
- Increased page size in DriverRideHistory for better data loading.
