+++ src/hooks/useUpdateDriverLocation.js
@@ -32,6 +32,7 @@ export default () => {
         const location = queueRef.current.shift();
 
         try {
+            console.log('calling driver location api with payload:', location);
             const response = updateDriverLocation(location.payload);
             await response.unwrap();
             lastProcessedRef.current = now;
