+++ src/navigation/GetNavigation.js
@@ -1,4 +1,4 @@
-import { lazy, Suspense } from 'react';
+import { lazy, Suspense, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { isUser, showErrorMessage } from '../util';
 import { isEmpty } from 'lodash';
@@ -16,45 +16,60 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 const DriverTabNavigator = lazy(() => (import('./driverTabNavigation')));
 const UserTabNavigator = lazy(() => (import('./userTabNavigation')));
 
-
 export const GetAuthRoutes = () => {
     const { driverInfo, userInfo } = useSelector(state => state.auth);
-    const isUserLogged = isUser()
-    const { logOut } = useLogout()
-    const { t } = useTranslation()
-    const roles = driverInfo?.DriverRoles || userInfo?.roles || [];
-    const isDriverLogged = roles.includes(USER_ROLES.DRIVER);
-    const isOwnerLogged = roles.includes(USER_ROLES.OWNER);
-    if (isDriverLogged) {
-        route = <Suspense fallback={<ActivityIndicator size="large" color="#0000ff" />}>
-            <DriverTabNavigator />
-        </Suspense>
-    } else if (isUserLogged) {
-        route = <Suspense fallback={<ActivityIndicator size="large" color="#0000ff" />}>
-            <UserTabNavigator />
-        </Suspense>
-    } else if (isOwnerLogged) {
-        route = <Suspense fallback={<ActivityIndicator size="large" color="#0000ff" />}>
-            <OwnerTabNavigator />
-        </Suspense>
-    } else {
-        logOut()
-        showErrorMessage(t('login_permission_error'))
-        route = <LoginNavigator />;
-    }
-    return route
-}
-export default () => {
-    const { access_token } = useSelector(state => state.auth);
-    useNotifications()
+    const { logOut } = useLogout();
+    const { t } = useTranslation();
 
-    let route = null
-    if (isEmpty(access_token)) {
-        route = <LoginNavigator />;
-    } else {
-        route = <SafeAreaView style={{ flex: 1, backgroundColor: COLORS.white }} edges={['top', 'left', 'right']}>
-            <GetAuthRoutes />
-         </SafeAreaView>
-    }
-    return route
+    const route = useMemo(() => {
+        const roles = driverInfo?.DriverRoles || userInfo?.roles || [];
+        const isDriverLogged = roles.includes(USER_ROLES.DRIVER);
+        const isOwnerLogged = roles.includes(USER_ROLES.OWNER);
+        const isUserLogged = roles.includes(USER_ROLES.USER);
+
+        if (isDriverLogged) {
+            return (
+                <Suspense fallback={<ActivityIndicator size="large" color="#0000ff" />}>
+                    <DriverTabNavigator />
+                </Suspense>
+            );
+        } else if (isUserLogged) {
+            return (
+                <Suspense fallback={<ActivityIndicator size="large" color="#0000ff" />}>
+                    <UserTabNavigator />
+                </Suspense>
+            );
+        } else if (isOwnerLogged) {
+            return (
+                <Suspense fallback={<ActivityIndicator size="large" color="#0000ff" />}>
+                    <OwnerTabNavigator />
+                </Suspense>
+            );
+        } else {
+            logOut();
+            showErrorMessage(t('login_permission_error'));
+            return <LoginNavigator />;
+        }
+    }, [driverInfo, userInfo, logOut, t]);
+
+    return route;
+};
+
+export default () => {
+    const access_token = useSelector(state => state.auth.access_token);
+    useNotifications();
+
+    const route = useMemo(() => {
+        if (isEmpty(access_token)) {
+            return <LoginNavigator />;
+        } else {
+            return (
+                <SafeAreaView style={{ flex: 1, backgroundColor: COLORS.white }} edges={['top', 'left', 'right']}>
+                    <GetAuthRoutes />
+                </SafeAreaView>
+            );
+        }
+    }, [access_token]);
+
+    return route;
 };
