+++ src/styles/RideMapStyles.js
@@ -1,0 +1,24 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    container: {
+        ...StyleSheet.absoluteFillObject,
+        justifyContent: 'flex-end',
+        alignItems: 'center',
+    },
+    map: {
+        ...StyleSheet.absoluteFillObject,
+    },
+    markerImageLarge: {
+        minHeight: 5,
+        minWidth: 5,
+        height: 40,
+        width: 40
+    },
+    markerImageSmall: {
+        minHeight: 5,
+        minWidth: 5,
+        height: 30,
+        width: 30
+    }
+});
